+++ changelog.md
@@ -1,13 +1,17 @@
 # Changelog
 
-## 0.3.0
+## v0.4.0
+
+* Add `error.ErrorChain`
+
+## v0.3.0
 
 * Change Queue.constructor(items), items is copy to internal queue instead of
   using directly.
 * Change Stack.constructor(items), items is copy to internal stack instead of
-  using directly. 
+  using directly.
 
-## 0.2.0
+## v0.2.0
 
 * Rename error.Useless to error.Closed
 * Add AsynQueue